переверстаны заметки сделан текст при отстувии юр лица сделала возможность добавление товара double передела кнопку плюса и начала делать цвет нижнего бара в разделе складов

+++ features/application-sections/profile-section/warehouse-feature/warehouse-impl/src/commonMain/kotlin/com/profile/profile/screens/main_refactor/screens/arrival_and_consumption_goods/model/StoreResponseArrivalAndConsumption.kt
@@ -41,7 +41,7 @@ data class ProductDataArrivalAndConsumption(
 
     val product_id: Int?,
 
-    val count: Int?,
+    val count: Double?,
 
 )
